requirements.txt
# --- Flask core ---
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-WTF==1.2.1
Flask-Mail==0.9.1
flask-cors
email-validator==2.2.0
WTForms==3.1.2
python-dotenv==1.0.1
Werkzeug==3.0.4
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5

# --- Database migration ---
alembic==1.14.0

# --- Images / Media ---
Pillow==11.0.0        # للتعامل مع الصور (قص – تغيير حجم – حفظ)
python-magic==0.4.27  # للتعرف على نوع الملف (مفيد للتحقق من الصور)

# --- PDF tools ---
PyPDF2==3.0.1         # لقراءة/معالجة ملفات PDF
reportlab==4.2.5      # لإنشاء ملفات PDF (تقارير – نشرات)
pdf2image==1.17.0     # لتحويل صفحات PDF إلى صور (يحتاج Poppler)
WeasyPrint==63.0      # لإنشاء ملفات PDF مباشرة بدون wkhtmltopdf
tinycss2==1.2.1       # تبع WeasyPrint لمعالجة CSS
cssselect2==0.7.0     # تبع WeasyPrint لمعالجة Selectors
html5lib==1.1         # لتفسير HTML داخليًا بشكل آمن
CairoSVG==2.7.1       # لتحويل شعارات/صور SVG إلى PDF تلقائيًا
xhtml2pdf==0.2.15     # بديل خفيف لتوليد PDF من HTML (خطة احتياطية)

# --- Utilities ---
requests==2.32.3      # لاستدعاءات HTTP (تكاملات – APIs)

# --- Security / Login (اختياري لكن مفيد) ---
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1

# --- IIS / Windows Deployment ---
wfastcgi              # تكامل Python مع IIS عبر FastCGI

# --- AI / Integrations ---
openai>=1.0.0
flask-cors

# --- Office / Documents ---
python-docx==0.8.11
